feat(bug fix): 설계, 보안, 버그 발생 가능한 메서드 및 클래스 수정(Binary, Participation, ChannelManage)
**
BinaryContent의 삭제 데이터 삭제가 정상적으로 되지 않았지만 그 예하에 있는 객체가 지워져 이후 삭제 되지않는 문제를 해결했습니다.
 Participation의 isOwner매서드가 Admin이 아닌경우 true를 반환했던 문제를 해결했습니다.
 ChannelManage에서 하위 계층 DTO를 참조해 설계상 오점이있던 문제를 해당 메서드를 하위 계층 Serv에 일임하며 오류를 해결했습니다.
**
