+++ src/main/java/com/sprint/mission/discodeit/application/service/ParticipationManagementService.java
@@ -3,5 +3,4 @@ package com.sprint.mission.discodeit.application.service;
 import java.util.UUID;
 
 public interface ParticipationManagementService {
-    int countNotReadChannelMessage(UUID channelId, UUID authorId);
 }
