+++ src/main/java/com/sprint/mission/discodeit/application/service/impl/AuthServiceImpl.java
@@ -1,5 +1,6 @@
 package com.sprint.mission.discodeit.application.service.impl;
 
+import com.sprint.mission.discodeit.application.service.AuthService;
 import com.sprint.mission.discodeit.channel.ChannelService;
 import com.sprint.mission.discodeit.message.channel.ChannelMessageService;
 import com.sprint.mission.discodeit.message.direct.DirectMessageService;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class AuthServiceImpl {
+public class AuthServiceImpl implements AuthService {
     private final UserService userService;
     private final ParticipationService participationService;
     private final DirectMessageService directMessageService;
