+++ src/main/java/com/sprint/mission/discodeit/application/service/impl/ChannelManagementServiceImpl.java
@@ -1,8 +1,9 @@
 package com.sprint.mission.discodeit.application.service.impl;
 
+import com.sprint.mission.discodeit.application.service.ChannelManagementService;
 import org.springframework.stereotype.Service;
 
 @Service
-public class ChannelManagementServiceImpl {
+public class ChannelManagementServiceImpl implements ChannelManagementService {
 
 }
